+++ acme-certify.cabal
@@ -13,10 +13,11 @@ cabal-version:       >=1.10
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Network.ACME.Encoding
+  exposed-modules:     Network.ACME, Network.ACME.Encoding
   build-depends:       base >= 4.7 && < 5,
                        cryptonite, aeson, bytestring, base64-bytestring, SHA,
-                       text, HsOpenSSL, time, email-validate
+                       mtl, text, HsOpenSSL, wreq, lens, lens-aeson, time,
+                       email-validate, pipes, directory, network-uri
   default-language:    Haskell2010
 
 executable acme-certify
@@ -25,9 +26,8 @@ executable acme-certify
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N -Wall
   build-depends:       base, acme-certify,
                        cryptonite, aeson, bytestring, base64-bytestring, SHA,
-                       text, HsOpenSSL, wreq, lens, lens-aeson,
-                       optparse-applicative, directory, mtl, time, pipes,
-                       email-validate, network-uri
+                       text, HsOpenSSL, optparse-applicative, time,
+                       email-validate, network-uri, directory
   default-language:    Haskell2010
 
 -- test-suite acme-certify-test
